feedview auto refresh when a weesh is added mapview with custom markers and custom description windows
:q

+++ UptoMobileIHM/YouWeesh.Mobile.IHM/YouWeesh.Mobile.IHM/App.cs
@@ -11,8 +11,8 @@ namespace YouWeesh.Mobile
 
     public class App : Application
     {
-		static public int ScreenWidth;
-		static public int ScreenHeight;
+		static public double ScreenWidth;
+		static public double ScreenHeight;
 
         public App()
         {
